requirements.txt
watchdog
gevent-websocket
gevent
ws4py
peewee
bcrypt
pyjwt
